refactor(domain): split state module into submodules
