+++ crates/luban_domain/src/state/attachments.rs
@@ -1,0 +1,24 @@
+#[derive(Clone, Copy, Debug, Eq, PartialEq, serde::Serialize, serde::Deserialize)]
+#[serde(rename_all = "snake_case")]
+pub enum AttachmentKind {
+    Image,
+    Text,
+    File,
+}
+
+#[derive(Clone, Debug, Eq, PartialEq, serde::Serialize, serde::Deserialize)]
+pub struct AttachmentRef {
+    pub id: String,
+    pub kind: AttachmentKind,
+    pub name: String,
+    pub extension: String,
+    pub mime: Option<String>,
+    pub byte_len: u64,
+}
+
+#[derive(Clone, Debug, Eq, PartialEq, serde::Serialize, serde::Deserialize)]
+pub struct ContextItem {
+    pub id: u64,
+    pub attachment: AttachmentRef,
+    pub created_at_unix_ms: u64,
+}
