+++ crates/luban_domain/src/state/ids.rs
@@ -1,0 +1,38 @@
+#[derive(Clone, Copy, Debug, Eq, PartialEq, Hash)]
+pub struct ProjectId(pub(crate) u64);
+
+impl ProjectId {
+    pub fn as_u64(self) -> u64 {
+        self.0
+    }
+
+    pub fn from_u64(id: u64) -> Self {
+        Self(id)
+    }
+}
+
+#[derive(Clone, Copy, Debug, Eq, PartialEq, Hash)]
+pub struct WorkspaceId(pub(crate) u64);
+
+impl WorkspaceId {
+    pub fn as_u64(self) -> u64 {
+        self.0
+    }
+
+    pub fn from_u64(id: u64) -> Self {
+        Self(id)
+    }
+}
+
+#[derive(Clone, Copy, Debug, Eq, PartialEq, Hash)]
+pub struct WorkspaceThreadId(pub(crate) u64);
+
+impl WorkspaceThreadId {
+    pub fn as_u64(self) -> u64 {
+        self.0
+    }
+
+    pub fn from_u64(id: u64) -> Self {
+        Self(id)
+    }
+}
